Refactor validator accounts backup to remove cli context dependency
